:recycle: refator: move area-repo to src dir
