+++ api/lib/infrastructure/repositories/campaign-participation-result-repository.js
@@ -2,7 +2,7 @@ import { CampaignParticipationResult } from '../../domain/models/CampaignPartici
 import * as campaignParticipationRepository from './campaign-participation-repository.js';
 import * as campaignRepository from './campaign-repository.js';
 import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
-import * as areaRepository from './area-repository.js';
+import * as areaRepository from '../../../src/shared/infrastructure/repositories/area-repository.js';
 import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as knowledgeElementRepository from './knowledge-element-repository.js';
